비밀번호 및 제목 넣고 API에 비빌번호 확인 코드 삽입.

+++ src/main/java/shop/betabeta/week03/controller/MemoController.java
@@ -6,8 +6,10 @@ import shop.betabeta.week03.domain.Memo;
 import shop.betabeta.week03.domain.MemoRepository;
 import shop.betabeta.week03.domain.MemoRequestDto;
 import shop.betabeta.week03.service.MemoService;
+import sun.security.util.Password;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @RestController
@@ -21,21 +23,33 @@ public class MemoController {
         Memo memo = new Memo(requestDto);
         return memoRepository.save(memo);
     }
+
     @GetMapping("/api/memos")
     public List<Memo> getMemos() {
         return memoRepository.findAllByOrderByModifiedAtDesc();
     }
 
     @DeleteMapping("/api/memos/{id}")
-    public Long deleteMemo(@PathVariable Long id) {
-        memoRepository.deleteById(id);
+    public Long deleteMemo(@PathVariable Long id, @RequestBody MemoRequestDto requestDto) {
+        Optional<Memo> memo = memoRepository.findById(id);
+        if (memo.get().getPassword().equals(requestDto.getPassword())) {
+            memoRepository.deleteById(id);
+        } else {
+            System.out.println("비밀번호 불일치");
+        }
         return id;
     }
 
     @PutMapping("/api/memos/{id}")
-    public Long updateMemo(@PathVariable Long id, @RequestBody MemoRequestDto requestDto){
-        memoService.update(id, requestDto);
+    public Long updateMemo(@PathVariable Long id, @RequestBody MemoRequestDto requestDto) {
+        Optional<Memo> memo = memoRepository.findById(id);
+        if (memo.get().getPassword().equals(requestDto.getPassword())) {
+            memoService.update(id, requestDto);
+        } else {
+            System.out.println("비밀번호 오류");
+        }
         return id;
+
     }
 
 }
